fixed string literal encoding issue (UTF-8 support)
MSVC doesn't assume that the source charset is UTF-8 (at least without
the BOM, which isn't required/recommended generally), so a flag has
been added to set the source/execution charsets to UTF-8.

Note that (at least MSVC) does not seem to care about the execution
charset for wide string literals (which is what is used in this case),
so strictly only the source charset matters in this case. That said,
it seems unwise to not have them match (and would break for non-wide
string literals in the future).

The reason the previous MSVC-specific solution worked (using a non-wide
string literal), despite the source/execution charset being wrong, is
that they were the _same_ (and no validation was enabled). As a result,
the string literal during execution contained exactly the bytes in the
source file, which is what the call to MultiByteToWideChar needed.